bundling cxf dependancies & refactor
