+++ components/inbound-endpoints/org.wso2.carbon.inbound.endpoint.ext.wsrm/src/main/java/org/wso2/carbon/inbound/endpoint/ext/wsrm/InboundRMServiceImpl.java
@@ -16,7 +16,7 @@
 * under the License.
 */
 
-package org.wso2.carbon.endpoint.ext.wsrm;
+package org.wso2.carbon.inbound.endpoint.ext.wsrm;
 
 /**
  * This is a dummy class to instantiate the RM Inbound Endpoint
